old-tests/submission/W3C/domhtml/HTMLIFrameElement08.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - IFRAME</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-iframe-scrolling" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<iframe height="80" scrolling="yes" src="/support/html/sample.html" width="550">IFRAME1</iframe>


<script>
var domtest = async_test("The scrolling IDL attribute of the frame element must reflect the respective content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vscrolling;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("iframe");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vscrolling = testNode.scrolling;

      assert_equals(vscrolling,"yes");
       
    domtest.done();
}));

</script>
<p>The scrolling IDL attribute of the frame element must reflect the respective content attribute of the same name.</p>
<div id="log"></div>
</body></html>
